Function reference

Image loaders

TestImages.testimage — Function

img = testimage(filename; download_only=false, [ops...])

Load test image that partially matches filename, the first match is used if there're more than one.

If download_only=true, the full filepath is returned. Any other keyword arguments ops will be passed to image IO backend through FileIO.load.

Example

julia> using TestImages
julia> img = testimage("cameraman.tif"); # fullname
julia> img = testimage("cameraman"); # without extension works
julia> img = testimage("c"); # with only partial name also works

Extended help

The following is a complete list of testimages, you can also check them at https://testimages.juliaimages.org/

"airplaneF16"

"autumn_leaves"

"barbara_color"

"barbara_gray_512"

"blobs"

"cameraman"

"chelsea"

"coffee"

"earth_apollo17"

"fabio_color_256"

"fabio_color_512"

"fabio_gray_256"

"fabio_gray_512"

"hela-cells"

"house"

"jetplane"

"lake_color"

"lake_gray"

"lena_color_256"

"lena_color_512"

"lena_gray_16bit"

"lena_gray_256"

"lena_gray_512"

"lighthouse"

"livingroom"

"m51"

"mandril_color"

"mandril_gray"

"mandrill"

"monarch_color"

"monarch_color_256"

"moonsurface"

"morphology_test_512"

"mountainstream"

"mri-stack"

"multi-channel-time-series.ome"

"peppers_color"

"peppers_gray"

"pirate"

"resolution_test_1920"

"resolution_test_512"

"simple_3d_ball"

"simple_3d_psf"

"sudoku"

"toucan"

"walkbridge"

"woman_blonde"

"woman_darkhair"

TestImages.shepp_logan — Function

phantom = shepp_logan(N, [M=N]; high_contrast=true)

Output the NxM Shepp-Logan phantom, which is a standard test image usually used for comparing image reconstruction algorithms in the field of computed tomography (CT) and magnetic resonance imaging (MRI).

If the argument M is omitted, the phantom is of size NxN. When setting the keyword argument high_constrast to false, the CT version [1] of the phantom is created. Otherwise, the high contrast MRI version [2] is calculated.

References

[1] Shepp, Lawrence A., and Benjamin F. Logan. "The Fourier reconstruction of a head section." IEEE Transactions on nuclear science 21.3 (1974): 21-43.

[2] Toft, Peter Aundal. "The Radon transform-theory and implementation." (1996): 201.

[3] Jain, Anil K. Fundamentals of digital image processing. Prentice-Hall, Inc., (1989): 439.

Helpers

TestImages.full_imagename — Function

fullname = full_imagename(shortname)

Get the first match of shortname in TestImages.remotefiles